chore: update deploy script

+++ scripts/stack-db-backup.sh
@@ -18,6 +18,12 @@ DATE=$(date +%Y-%m-%d_%H-%M)
 BACKUP_PATH="/mnt/volume_rhea_postgres/backups/db.sql"
 UPLOAD_PATH="s3://$BUCKET/rhea/backups/db-$DATE.sql"
 
+# Sanity check that the backup file exists and is not empty
+if [ ! -s "$BACKUP_PATH" ]; then
+  echo "Backup file does not exist or is empty."
+  exit 1
+fi
+
 # Upload the backup to DigitalOcean Spaces
 aws s3 cp $BACKUP_PATH $UPLOAD_PATH --endpoint-url https://fra1.digitaloceanspaces.com
 echo "Successfully uploaded the latest backup to $UPLOAD_PATH"
